address comments and rework migration
