+++ fence/oidc/grants/__init__.py
@@ -1,4 +1,4 @@
 from fence.oidc.grants.implicit_grant import ImplicitGrant
-from fence.oidc.grants.oidc_code_grant import OpenIDCodeGrant
+from fence.oidc.grants.oidc_code_grant import AuthorizationCodeGrant
 from fence.oidc.grants.refresh_token_grant import RefreshTokenGrant
 from fence.oidc.grants.client_credentials_grant import ClientCredentialsGrant
