+++ pyproject.toml
@@ -13,7 +13,7 @@ include = [
 [tool.poetry.dependencies]
 python = ">=3.9,<4.0.0"
 alembic = "^1.7.7"
-authlib = ">=1.1.0"
+authlib = "*" # let authutils decide which version to use
 authutils = {git = "https://github.com/uc-cdis/authutils", rev = "feat/authlib"}
 bcrypt = "^3.1.4"
 boto3 = "~1.9.91"
