+++ tests/rfc6749/test_oauth2.py
@@ -28,7 +28,8 @@ def test_oauth2_authorize_incorrect_scope(oauth_test_client, method):
     auth_response = oauth_test_client.authorize(
         method=method, data=data, do_asserts=False
     )
-    assert auth_response.response.status_code == 400
+    # Check the status code is not a redirect code 3xx
+    assert str(auth_response.response.status_code)[0] != "3"
 
 
 @pytest.mark.parametrize("method", ["GET", "POST"])
